Compress the program under 1 package
