+++ src/main/java/mon/Mon.java
@@ -1,0 +1,70 @@
+package mon;
+
+import java.util.ArrayList;
+
+import mon.command.Command;
+import mon.command.Parser;
+import mon.exception.MonException;
+import mon.storage.Storage;
+import mon.task.Task;
+import mon.task.TaskList;
+import mon.ui.Ui;
+
+public class Mon {
+    private static final String FILE_PATH = "data/mon.txt";
+
+    private final Storage storage;
+    private final TaskList taskList;
+    private final Ui ui;
+
+    public Mon() {
+        this.ui = new Ui();
+        this.storage = new Storage(FILE_PATH);
+        TaskList tempTaskList;
+        try {
+            ArrayList<Task> tasks = storage.loadTasks();
+            tempTaskList = new TaskList(tasks);
+        } catch (MonException e) {
+            ui.showError("Error loading tasks: " + e.getMessage());
+            tempTaskList = new TaskList();
+        }
+        this.taskList = tempTaskList;
+    }
+
+    public static void main(String[] args) {
+        Mon mon = new Mon();
+        mon.run();
+    }
+
+    public void run() {
+        ui.showWelcome();
+
+        boolean shouldExit = false;
+        
+        while (ui.hasNextLine() && !shouldExit) {
+            String input = ui.readCommand();
+            shouldExit = handleInput(input);
+        }
+        
+        ui.close();
+    }
+
+    private boolean handleInput(String input) {
+        try {
+            Command command = Parser.parse(input);
+            String response = command.execute(taskList, storage);
+            
+            ui.showMessage(response);
+            
+            // Save tasks after every command that modifies the task list
+            if (!command.isExit()) {
+                storage.saveTasks(taskList.getTaskList());
+            }
+            
+            return command.isExit();
+        } catch (Exception e) {
+            ui.showError(e.getMessage());
+            return false;
+        }
+    }
+}
