+++ src/main/java/mon/command/ExitCommand.java
@@ -1,0 +1,21 @@
+package mon.command;
+
+import mon.storage.Storage;
+import mon.task.TaskList;
+
+/**
+ * Command to exit the application.
+ */
+public class ExitCommand extends Command {
+    private static final String GOODBYE_MESSAGE = "Mon: See you again!";
+    
+    @Override
+    public String execute(TaskList taskList, Storage storage) throws Exception {
+        return GOODBYE_MESSAGE;
+    }
+    
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
