+++ src/main/java/mon/task/TaskList.java
@@ -1,0 +1,35 @@
+package mon.task;
+
+import java.util.ArrayList;
+
+public class TaskList {
+    private ArrayList<Task> tasks;
+
+    public TaskList(ArrayList<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public TaskList() {
+        this.tasks = new ArrayList<>();
+    }
+
+    public ArrayList<Task> getTaskList() {
+        return tasks;
+    }
+
+    public Task getTask(int index) {
+        return tasks.get(index);
+    }
+
+    public void addTask(Task task) {
+        tasks.add(task);
+    }
+
+    public void removeTask(int index) {
+        tasks.remove(index);
+    }
+
+    public int size() {
+        return tasks.size();
+    }
+}
